add statistics

+++ domain/statistics/src/commonMain/kotlin/models/ExerciseVolume.kt
@@ -4,5 +4,5 @@ public data class ExerciseVolume(
     val createdAt: String,
     val exerciseExampleId: String,
     val id: String,
-    val volume: Int
+    val volume: Double
 )
